fix: eliminate seperate code artifact component and unify into one.
